[artf4916] This revision includes the first version of the SvnEdge Bonjour client that can be integrated in other projects. It includes the dependencies to the jmDNS and provides APIs to "observe" and discover svnedge servers in a local network. The project is called "SvnEdge Discovery API" and requires an implementation of the jmDNS package. The package includes an example that can be excecuted directly from running "ant run".
The main client is called SvnEdgeBonjourClient, which offers users to select a service using the enum SvnEdgeServiceType. As on the Bonjour protocol, the user can get the list of expected keys SvnEdgeServiceKey related to the type. As the current implementation, the CSVN and HTTP service types are implemented by SvnEdgeCsvnServiceKey and SvnEdgeHttpServiceKey. In this way, the SvnEdgeBonjourClient uses the type to filter the packets received from the jmDNS service and broadcasts the events "server is running" and "server stopped" to a list of SvnEdgeServerObserver, which needs to be added as an observer. Once one of those events is captured and translated from the mDNS service, the information related to the packet is wrapped as an instance of SvnEdgeServerInfo containing the major data information such as IP address, URL, and a map of SvnEdgeServiceKey->String of the associated keys.

The implementation suggests the use of the package "com.collabnet.svnedge.mdns". 

TODO: Service naming, versioning, and other types should be added here, and reused by the SvnEdge main project.
TODO: Integrate with the client project, removing the jmDNS dependencies.
TODO: Integrate with the SvnEdge main project, removing the jmDNS dependencies.

* .classpath
* .project
- The Eclipse project definition.

* dist/svnedge-discovery-api-1.0.0.jar
- The distribution jar of the API. It can be copied to other SvnEdge projects by running "ant update-distro".

* lib/jmdns.jar
- The current implementation of the jmdns.jar copied from the "client" project.

* SvnEdgeCsvnServiceKey.java
- The enum for the keys of the service type "CSVN" currently published by SvnEdge.

* SvnEdgeHttpServiceKey.java
- The enum for the keys of the service type "HTTP" currently published by SvnEdge.

* SvnEdgeServiceKey.java
- The marker interface used by the SvnEdgeServiceType.

* SvnEdgeServiceType.java
- The implementation of the service type.

* SvnEdgeBonjourClient.java
- The implementation of the client to be used by other projects.
- This class includes a driver to run in the command line. It selects the existing service types and requires the user to choose the value using the keyboard.

* SvnEdgeServerObserver.java
- The observer interface that lists the events used by the client. It proxies the jmDNS events to the observers when an SvnEdge server is found running or when one stops.

* SvnEdgeServerInfo.java
- The data wrapper from the jmDNS service Info. Users to the API can get all the packet values from this class.

* build.xml
- Contains the regular target "compile", "jar" and "run". The target "run" is the default. 
- Offers the task "update-distros" used to copy the current version to the projects "Discovery client" and the main "SvnEdge" project.


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@1264 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f